Added order realisation and fixed controllers
fixed status codes in controllers
added order realisation

+++ ServiceHub/src/main/java/org/servicehub/service/ServicehubUserService.java
@@ -75,6 +75,9 @@ public class ServicehubUserService {
 
     @Transactional
     public void remove(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException("User " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
 }
